integrate gil server and test
